+++ src/DoubanAwardsCollector/JsonDefaults.cs
@@ -1,3 +1,4 @@
+using System.Text.Encodings.Web;
 using System.Text.Json;
 
 namespace DoubanAwardsCollector;
@@ -12,6 +13,7 @@ internal static class JsonDefaults
     public static JsonSerializerOptions Write { get; } = new()
     {
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+        Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
         WriteIndented = true
     };
 }